move invalid Text mixins
