+++ invalid/main/java/org/spongepowered/common/mixin/invalid/core/api/text/LiteralTextMixin.java
@@ -22,23 +22,23 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.spongepowered.common.mixin.core.api.text;
+package org.spongepowered.common.mixin.invalid.core.api.text;
 
-import net.minecraft.util.text.SelectorTextComponent;
+import net.minecraft.util.text.StringTextComponent;
 import net.minecraft.util.text.TextComponent;
-import org.spongepowered.api.text.SelectorText;
-import org.spongepowered.api.text.selector.Selector;
+import org.spongepowered.api.text.LiteralText;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 
-@Mixin(value = SelectorText.class, remap = false)
-public abstract class SelectorTextMixin extends TextMixin {
+@Mixin(value = LiteralText.class, remap = false)
+public abstract class LiteralTextMixin extends TextMixin {
 
-    @Shadow @Final Selector selector;
+    @Shadow @Final String content;
 
     @Override
     protected TextComponent createComponent() {
-        return new SelectorTextComponent(this.selector.toPlain());
+        return new StringTextComponent(this.content);
     }
+
 }
