+++ invalid/main/java/org/spongepowered/common/mixin/invalid/core/api/text/TranslatableTextMixin.java
@@ -22,23 +22,40 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.spongepowered.common.mixin.core.api.text;
+package org.spongepowered.common.mixin.invalid.core.api.text;
 
-import net.minecraft.util.text.StringTextComponent;
+import com.google.common.collect.ImmutableList;
 import net.minecraft.util.text.TextComponent;
-import org.spongepowered.api.text.LiteralText;
+import net.minecraft.util.text.TranslationTextComponent;
+import org.spongepowered.api.text.TranslatableText;
+import org.spongepowered.api.text.translation.Translation;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.common.bridge.api.text.TextBridge;
 
-@Mixin(value = LiteralText.class, remap = false)
-public abstract class LiteralTextMixin extends TextMixin {
+@Mixin(value = TranslatableText.class, remap = false)
+public abstract class TranslatableTextMixin extends TextMixin {
 
-    @Shadow @Final String content;
+    @Shadow @Final Translation translation;
+    @Shadow @Final ImmutableList<Object> arguments;
 
     @Override
     protected TextComponent createComponent() {
-        return new StringTextComponent(this.content);
+        return new TranslationTextComponent(this.translation.getId(), this.unwrapArguments(this.arguments));
+    }
+
+    private Object[] unwrapArguments(final ImmutableList<Object> args) {
+        final Object[] result = new Object[args.size()];
+        for (int i = 0; i < args.size(); i++) {
+            final Object arg = args.get(i);
+            if (arg instanceof TextBridge) {
+                result[i] = ((TextBridge) arg).bridge$toComponent();
+            } else {
+                result[i] = arg;
+            }
+        }
+        return result;
     }
 
 }
